+++ src/TestEngine/nunit.engine.core/Helpers/XmlHelper.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2010-2014 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -32,7 +32,7 @@ namespace System.Runtime.CompilerServices
     sealed class ExtensionAttribute : Attribute { }
 }
 
-namespace NUnit
+namespace NUnit.Engine.Helpers
 {
     /// <summary>
     /// XmlHelper provides static methods for basic XML operations.
